ref: split large index page function into multiple functions
feat: changed buttons into links in greeter on index page

feat: removed dependacy for react-tiny-popover

feat: updated packages

feat: added turbopack for dev run
