+++ frontend/src/app/globals.css
@@ -50,7 +50,3 @@ main {
   align-items: center;
   align-items: center;
 }
-
-.none {
-  *: none;
-}
