--- frontend/src/components/Providers.tsx
@@ -1,27 +1,0 @@
-'use client'
-
-import { useState } from "react";
-import { ThemeProvider } from "@/components/ThemeProvider";
-import { 
-  QueryClient,
-  QueryClientProvider
-} from "@tanstack/react-query";
-
-
-
-export default function Providers({ children }: Readonly<{ children: React.ReactNode }>) {
-  const [queryClient] = useState(() => new QueryClient())
-
-  return (
-    <QueryClientProvider client={queryClient}>
-      <ThemeProvider
-        attribute="class"
-        defaultTheme="system"
-        enableSystem
-        disableTransitionOnChange
-      >
-        {children}
-      </ThemeProvider>
-    </QueryClientProvider>
-  )
-}
