+++ frontend/src/components/core/Footer.tsx
