+++ frontend/src/components/home/types.tsx
@@ -1,0 +1,7 @@
+type FeatureItem = {
+  readonly img: { src: string; alt: string };
+  readonly title: string;
+  readonly detail: string;
+};
+
+export default FeatureItem;
